Split storm_drainage.rs (748 lines) into subdirectory modules
Part of #1389

Split the monolithic storm_drainage.rs into a subdirectory with:
- mod.rs (29 lines): module doc, re-exports, plugin definition
- types.rs (89 lines): constants, StormDrainageType enum, component, state resource
- systems.rs (169 lines): update_storm_drainage system
- tests.rs (477 lines): all unit tests

All pub items re-exported from mod.rs so external imports are unchanged.
